+++ src/agent/agent.go
@@ -25,6 +25,7 @@ type Agent struct {
 	Client     *api.Client
 	datacenter string
 	ipAddr     string
+	name       string
 }
 
 // CreateAgents creates an Agent structure for every Agent member of the LAN cluster
@@ -44,19 +45,20 @@ func CreateAgents(client *api.Client, i *integration.Integration, args *args.Arg
 	agents = make([]*Agent, 0, len(members))
 	for _, member := range members {
 
-		entity, err := i.Entity(member.Name, "agent")
+		memberNameIDAttr := integration.NewIDAttribute("co-agent", member.Name)
+		entity, err := i.Entity(fmt.Sprintf("%s:%d", member.Addr, member.Port), "co-agent", memberNameIDAttr)
 		if err != nil {
 			log.Error("Error creating entity for Agent '%s': %s", member.Name, err.Error())
 			continue
 		}
 
-		client, err = api.NewClient(args.CreateAPIConfig(member.Name))
+		client, err = api.NewClient(args.CreateAPIConfig(member.Addr))
 		if err != nil {
 			log.Error("Error creating client for Agent '%s': %s", member.Name, err.Error())
 			continue
 		}
 
-		agent := NewAgent(client, entity, member.Addr, member.Tags["dc"])
+		agent := NewAgent(client, entity, member.Name, member.Addr, member.Tags["dc"])
 		agents = append(agents, agent)
 
 		// we need to identify the leader to collect catalog
@@ -71,12 +73,13 @@ func CreateAgents(client *api.Client, i *integration.Integration, args *args.Arg
 }
 
 // NewAgent creates a new agent from the given client and Entity
-func NewAgent(client *api.Client, entity *integration.Entity, ipAddr, datacenter string) *Agent {
+func NewAgent(client *api.Client, entity *integration.Entity, name, ipAddr, datacenter string) *Agent {
 	return &Agent{
 		Client:     client,
 		entity:     entity,
 		ipAddr:     ipAddr,
 		datacenter: datacenter,
+		name:       name,
 	}
 }
 
@@ -134,12 +137,12 @@ func (a *Agent) collectLatencyMetrics(metricSet *metric.Set) error {
 	}
 
 	if len(nodes) == 1 {
-		return errors.New("cluster only contains 1 node")
+		return errors.New("could not collect latency metrics because the cluster only contains 1 node")
 	}
 
-	agentNode := findNode(a.entity.Metadata.Name, nodes)
+	agentNode := findNode(a.name, nodes)
 	if agentNode == nil {
-		return errors.New("could not find node for agent")
+		return fmt.Errorf("could not collect latency metrics because the node with name %s could not be found be found in the list of nodes: %v", a.name, nodes)
 	}
 
 	// calculate and populate metrics
@@ -149,8 +152,8 @@ func (a *Agent) collectLatencyMetrics(metricSet *metric.Set) error {
 	return nil
 }
 
-// Name returns the entity name of the agent
-func (a *Agent) Name() string {
+// HostPort returns the host:port of the agent
+func (a *Agent) HostPort() string {
 	return a.entity.Metadata.Name
 }