+++ src/datacenter/datacenter_test.go
@@ -68,7 +68,7 @@ func TestNewDatacenter_Normal(t *testing.T) {
 
 	if out.entity.Metadata.Name != "dev-uss" {
 		t.Fatalf("Entity was not named correctly %s", out.entity.Metadata.Name)
-	} else if out.entity.Metadata.Namespace != "datacenter" {
+	} else if out.entity.Metadata.Namespace != "co-datacenter" {
 		t.Fatalf("Entity has wrong namespace %s", out.entity.Metadata.Namespace)
 	}
 
@@ -213,7 +213,7 @@ func Test_Datacenter_CollectMetrics_Full(t *testing.T) {
 
 	c := &Datacenter{
 		entity: dcEntity,
-		leader: agent.NewAgent(client, agentEntity, "", ""),
+		leader: agent.NewAgent(client, agentEntity, "", "", ""),
 	}
 
 	setMetricMuxes(mux)
@@ -274,7 +274,7 @@ func Test_Datacenter_CollectMetrics_All_Endpoint_Fails(t *testing.T) {
 
 	c := &Datacenter{
 		entity: dcEntity,
-		leader: agent.NewAgent(client, agentEntity, "", ""),
+		leader: agent.NewAgent(client, agentEntity, "", "", ""),
 	}
 
 	expected := map[string]interface{}{