+++ CHANGELOG.md
@@ -5,8 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## 1.0.2 - 2019-10-23
+## 2.0.3 - 2019-10-23
+### Added
 - Windows installer packaging
+
+## 2.0.2 - 2019-07-17
+### Fixed
+- Use agent name for latency calculations
+
+## 2.0.0 - 2019-04-25
+### Changed
+- Prefixed namespaces for uniqueness
+- Updated SDK
+- Added ID attributes
+
+## 1.1.1 - 2019-04-17
+### Added
+- Use address rather than name to connect
+
+## 1.1.0 - 2019-04-08
+### Added
+- Local-only collection option
+>>>>>>> master
 
 ## 1.0.1 - 2019-03-19
 ### Fixed